Transactions endpoint fixed

+++ src/main/java/com/yummy/transaction/model/TransactionItem.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class TransactionItem {
 
-    private int offerId;
+    private long offerId;
     private int count;
 
 }
